requirement.txt
torch
torchvision
pytorch-lightning 
tensorboard

# numpy
PyYAML
rich
opencv-python

matplotlib
albumentations